chore: add infos to package.json and refactor electron - types
