+++ src/main/database/transactions.ts
@@ -1,4 +1,4 @@
-import type { SqliteDatabase } from './types'
+import type { SqliteDatabase } from './types/database'
 
 export function runInTransaction(db: SqliteDatabase, fn: () => void) {
   if ('transaction' in db && typeof db.transaction === 'function') {
