+++ vitest.config.ts
@@ -8,7 +8,6 @@ const dirname =
   typeof __dirname !== 'undefined' ? __dirname : path.dirname(fileURLToPath(import.meta.url))
 const includeStorybookProject = process.env.VITEST_STORYBOOK === 'true'
 
-// More info at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon
 export default defineConfig({
   plugins: [vue()],
   optimizeDeps: {
@@ -115,10 +114,8 @@ export default defineConfig({
       ...(includeStorybookProject
         ? [
             {
-              extends: true,
+              extends: true as const,
               plugins: [
-                // The plugin will run tests for the stories defined in your Storybook config
-                // See options at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon#storybooktest
                 storybookTest({
                   configDir: path.join(dirname, '.storybook')
                 })
@@ -135,7 +132,7 @@ export default defineConfig({
                   provider: playwright({}),
                   instances: [
                     {
-                      browser: 'chromium'
+                      browser: 'chromium' as const
                     }
                   ]
                 }
